requirements.txt
flask
requests
python-dotenv
flask-socketio
